чистит мейн от лишнего кода
