+++ js/render-thumbnails.js
@@ -1,4 +1,3 @@
-
 // Создаёт DOM-элемент для одной фотографии
 
 function createThumbnail({ url, description, likes, comments }, template) {
@@ -31,5 +30,4 @@ function renderThumbnails(photos) {
   const thumbnails = photos.map((photo) => createThumbnail(photo, template));
   appendThumbnails(thumbnails, picturesContainer);
 }
-
 export { renderThumbnails };
